Little additionals to PMM init code
Buddy alloc library integrated

+++ kernel/src/memory_management/mod.rs
@@ -1,7 +1,7 @@
 mod physical_memory_manager;
 
 /// 4KB
-const PAGE_SIZE: u64 = 4096;
+const PAGE_SIZE: usize = 4096;
 
 /// Inits Physical Memory Manager and Virtual Memory Manager
 pub fn init(boot_info: &bootloader_api::BootInfo) {
